+++ src/component/collapse.js
@@ -1,0 +1,21 @@
+import React, {PropTypes} from 'react';
+import className from 'classnames';
+
+class Collapse extends React.Component {
+    render() {
+        const {children, ...rest} = this.props;
+        return (
+            <div {...rest} className={className('gm-collapse', this.props.className, {
+                'in': this.props.in
+            })}>
+                {children}
+            </div>
+        );
+    }
+}
+
+Collapse.propTypes = {
+    in: PropTypes.bool.isRequired
+};
+
+export default Collapse;