+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 移除`react-boostrap`依赖
 
-### api改动点
-- dialog 的 bsSize => size
+### 兼容性改动
+- Dialog 的 bsSize => size
 - app 要预埋 <LayoutRoot/>
+- AdvanceSelect 用到的类 gm-dropdown 改成 gm--dropdown
+
+### 新增
+- Modal
+- Collapse
+- DropDown